fix: move gen/ out of api/ to avoid Vercel scanning generated Go files
Vercel scans ALL .go files recursively in api/ as potential serverless
functions. Committed generated files in api/gen/go/ triggered:
  "Could not find an exported function in api/gen/go/sup/supconnect/..."

Fix: move generated output to gen/ at repo root (outside api/):
- buf.gen.yaml: out: gen/go and out: gen/python
- go_package in proto files updated to github.com/maybecake/stacks/gen/go/...
- api/yo.go imports updated accordingly
- api/sup.py path updated (../gen/python/sup relative to api/)
