+++ api/sup.py
@@ -5,7 +5,7 @@ from http.server import BaseHTTPRequestHandler
 from google.protobuf import json_format
 
 # Add generated proto path at runtime (committed to repo, generated by buf generate)
-_gen_path = os.path.join(os.path.dirname(__file__), "gen", "python", "sup")
+_gen_path = os.path.join(os.path.dirname(__file__), "..", "gen", "python", "sup")
 if _gen_path not in sys.path:
     sys.path.insert(0, _gen_path)
